Enhance API error handling and configuration checks
- Added warnings for missing configuration keys in the Cloudflare Turnstile token verification.
- Introduced timeout signals for fetch requests in various API routes to improve reliability.
- Updated notification services to ensure site URL is configured before sending notifications.
- Refactored image asset deletion logic for better clarity and error handling.
- Improved cache management and error handling in various utility functions.
